Moved the platform specific config creation to a separate function
